+++ scripts/renewdb_saving_data.sh
@@ -1,21 +1,21 @@
 #!/bin/bash
 
-if [[ $PWD == *"scripts"* ]]
-then
+export start_pwd=$PWD
+while [[ $PWD == *"scripts"* ]]; do
     cd ..
+done
+
+source ./scripts/auxiliary/prepare_launch.sh
+
+if $(password_check); then
+    pg_dump -Fc -Z0 --data-only -U mib_api metalinblood > ./mibdb_data.pgdump
+
+    dropdb -U mib_api metalinblood
+    createdb -U mib_api metalinblood
+    psql metalinblood -U mib_api < ./sql/mibdb.sql
+
+    pg_restore -U mib_api --data-only -d metalinblood ./sql/wkdb_data.pgdump
+    rm -f ./wkdb_data.pgdump
 fi
 
-if [ -z ${PGPASSWORD+x} ]; then
-    export PGPASSWORD
-    echo "Enter the PostgreSQL password:"
-    read -s PGPASSWORD
-fi
-
-pg_dump -Fc -Z0 --data-only -U mib_api metalinblood > ./mibdb_data.pgdump
-
-dropdb -U mib_api metalinblood
-createdb -U mib_api metalinblood
-psql metalinblood -U mib_api < ./sql/mibdb.sql
-
-pg_restore -U mib_api --data-only -d metalinblood ./sql/wkdb_data.pgdump
-rm -f ./wkdb_data.pgdump
+cd $start_pwd
